+++ controllers/contacts/getContactById.js
@@ -1,0 +1,16 @@
+const contacts = require("../../models/contacts");
+const { RequestError } = require("../../helpers");
+
+const getContactById = async (req, res, next) => {
+  const { contactId } = req.params;
+  const result = await contacts.getContactById(contactId);
+
+  if (!result) {
+    throw RequestError(404, "Not found");
+  }
+
+  res.json(result);
+  res.status(200);
+};
+
+module.exports = getContactById;
