+++ helpers/index.js
@@ -1,5 +1,7 @@
 const RequestError = require("./RequestError");
+const ctrlWrapper = require("./ctrlWrapper");
 
 module.exports = {
-    RequestError,
-}
+  RequestError,
+  ctrlWrapper,
+};
