+++ models/contacts.json
@@ -7,9 +7,9 @@
   },
   {
     "id": "2",
-    "name": "Chaim Lewis",
-    "email": "dui.in@egetlacus.ca",
-    "phone": "(294) 840-6685"
+    "name": "Jeck kjijie",
+    "email": "mattis.Cras@nonenimMauris.net",
+    "phone": "(542) 451-7038"
   },
   {
     "id": "3",
@@ -64,5 +64,17 @@
     "name": "Resdsen Hedfy",
     "email": "pharetra.ut@ditum.co.uk",
     "phone": "(715) 568-5710"
+  },
+  {
+    "id": "5b930110-55ac-11ed-bfe3-5d0381976349",
+    "name": "Jeck kjijie",
+    "email": "mattis.Cras@nonenimMauris.net",
+    "phone": "(542) 451-7038"
+  },
+  {
+    "id": "46c87550-55b4-11ed-bfe9-d719674451e4",
+    "name": "Jecsdk kjsdijie",
+    "email": "mattis.Cras@nonenimMauris.net",
+    "phone": "(542) 451-7038"
   }
 ]
